requirements.txt
# Gradio
gradio

# OpenAI
openai

# Hugging Face Transformers
transformers

# Pillow (PIL)
Pillow

#crewai multiagent
crewai

#langchain
langchain
langchain-ollama

#request for URL fetching
requests

#include the variable from .env
python-dotenv

# Data ingestion scraping
selenium

# PyTorch
torch

# Additional dependencies for NLP and data handling
numpy
pandas
scikit-learn
matplotlib
dspy
# Hugging Face Datasets (optional, if needed)
datasets
# For image processing (if needed)
opencv-python
